fix(bridge): address Oracle review — split test files, add edge case coverage
- Split opencode_api_test.dart (304 lines) into opencode_api_test.dart (101 lines) and opencode_api_http_test.dart (207 lines)
- Split session_metadata_generator_test.dart (307 lines) into session_metadata_generator_test.dart (208 lines) and session_metadata_generator_failure_test.dart (69 lines)
- Extracted shared test helpers to session_metadata_generator_test_helpers.dart (96 lines)
- Added test for embedded JSON extraction from trailing text
- Added test for empty provider ID (model string starts with /)
- Added test for empty model ID (model string ends with /)
- Added test for whitespace-only text parts being skipped in create_session_handler
- All files now ≤250 lines per file size policy
- All tests pass, analysis clean
